+++ tests/unit/demography_two/conftest.py
@@ -88,3 +88,25 @@ def rtmodel_data():
     rdata_arrays = {k: np.reshape(v, (3, 6)).T for k, v in rdata.items()}
 
     return rdata_arrays
+
+
+@pytest.fixture
+def fixture_cohorts_and_allometry():
+    """A fixture providing a simple community."""
+
+    from pyrealm.demography_two.cohorts import Cohorts
+    from pyrealm.demography_two.flora import Flora
+    from pyrealm.demography_two.tmodel import StemAllometry
+
+    # A simple community containing one sample stem, with an initial crown gap fraction
+    # of zero.
+    flora = Flora(name=["test"], f_g=[0.0])
+    cohorts = Cohorts(
+        pft_name=np.array(["test"] * 4),
+        dbh_value=np.array([0.2, 0.4, 0.6, 0.8]),
+        n_individuals=np.array([1] * 4),
+        flora=flora,
+    )
+    allometry = StemAllometry(cohorts=cohorts)
+
+    return cohorts, allometry
